Time: 12 ms (10.23%), Space: 41.7 MB (7.01%) - LeetHub

+++ 0069-sqrtx/0069-sqrtx.java
@@ -12,7 +12,7 @@ class Solution {
 
             if(sq == x){
                 return mid;
-            }else if(sq < (long)x){
+            }else if(sq <x){
                 l = mid+1;
             }else{
                 r=mid-1;
